requirements.txt
dashscope>=1.10.0
python-dotenv>=0.19.0
requests>=2.31.0
rich>=13.0.0
typer>=0.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0 